+++ app/src/main/java/com/sopt/clody/presentation/ui/home/calendar/component/MonthlyItem.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.unit.dp
 import com.sopt.clody.data.remote.dto.response.MonthlyCalendarResponseDto
-import com.sopt.clody.presentation.ui.home.model.CalendarDate
+import com.sopt.clody.presentation.ui.home.calendar.model.CalendarDate
 import com.sopt.clody.presentation.utils.amplitude.AmplitudeConstraints
 import com.sopt.clody.presentation.utils.amplitude.AmplitudeUtils
 import kotlinx.datetime.DayOfWeek
